Method added into the controller to receive products

+++ Web Scraper/project/CuokaWebScraperRESTServer/src/main/java/es/sidelab/cuokawebscraperrestserver/beans/BeanProduct.java
@@ -19,17 +19,17 @@ public class BeanProduct
     private String name;
     private String shop;
     private String section;
-    private float price;
+    private double price;
     
     public BeanProduct() {}
     
     public String getName() { return this.name; }
     public String getShop() { return this.shop; }
     public String getSection() { return this.section; }
-    public float getPrice() { return this.price; }
+    public double getPrice() { return this.price; }
     
     public void setName( String name ) { this.name = name; }
     public void setShop( String shop ) { this.shop = shop; }
     public void setSection( String section ) { this.section = section; }
-    public void setPrice( float price ) { this.price = price; }
+    public void setPrice( double price ) { this.price = price; }
 }
